libeos-update-server: Move EosUpdateServerRepo into a utility library
This will make it easier to unit test in future. (The unit tests don’t
currently exist though).

This removes the installed eos-repo-server.h header, since installing it
was fairly pointless — there are no known users of this public API, and
it is going to change shortly.


https://phabricator.endlessm.com/T16551
